% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{This script includes different functions for data cleaning and manipulation
Clean NIRS Data Using Rolling Z-Scores}
\usage{
clean(path, winsize, vars)
}
\arguments{
\item{path}{Character. The file path where the CSV files are located.}

\item{winsize}{Integer. The window size to be used for calculating the rolling z-scores.}

\item{vars}{Character vector. The names of the variables (columns) in the data frames to be cleaned.}
}
\value{
A list of cleaned data frames.
}
\description{
This function reads multiple CSV files containing NIRS data, applies a rolling z-score
method to identify and remove outliers, and returns a list of cleaned data frames.
Filename may include the ID in the 7 first instances of the file name (i.e., 001-001)
while the first 3 instances may indicate the population (i.e., 001).
Columns are added in each file indicating ID and population.
}
\examples{
# Example usage:
path <- "path/to/csv/files"
winsize <- 50
vars <- c("var1", "var2", "var3")
cleaned_data <- clean(path, winsize, vars)
}
