% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_smooth.R
\name{filter_smooth}
\alias{filter_smooth}
\title{Filter Data Using Savitzky-Golay Filter}
\usage{
filter_smooth(path_or_list, vars)
}
\arguments{
\item{path_or_list}{Path or list. If path, the file path where the CSV files are located.
If list, a list of data frames.}

\item{vars}{Character vector. The names of the variables (columns) in the data frames to be filtered.}
}
\value{
A list of filtered data frames.
}
\description{
This function applies the Savitzky-Golay filter to smooth the data in specified variables
of data frames stored in a list or CSV files from a path.
}
\examples{
# Example usage with file path:
path <- "path/to/csv/files"
vars <- c("var1", "var2", "var3")
filtered_data <- filter_smooth(path, vars)

# Example usage with list of data frames:
data_list <- list(data1, data2, data3)
vars <- c("var1", "var2")
filtered_smoothed <- filter_smooth(data_list, vars)
}
\references{
Savitzky A, Golay MJE (1964). "Smoothing and Differentiation of Data by Simplified Least Squares Procedures."
Analytical Chemistry, 36(8), 1627-1639. doi: 10.1021/ac60214a047
}
\seealso{
\code{\link{sgolayfilt}}
}
