% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_decay.R
\name{single_decay}
\alias{single_decay}
\title{Single Exponential Decay Fitting}
\usage{
single_decay(
  path_or_list,
  vars,
  event_column,
  start_event,
  time_column,
  transient_phase = NULL,
  eyeball_data = "yes",
  plot_fitted,
  decay_start,
  y_Bas_user,
  A_p_user,
  T_Dp_user,
  tau_p_user
)
}
\arguments{
\item{path_or_list}{A list of data frames or a directory path containing CSV files to be processed.}

\item{vars}{A vector of a single column name in the data frames to which the decay model should be applied.}

\item{event_column}{The column name (as a string) in the data frames that contains the event markers.}

\item{start_event}{The value in the event column that marks the start of the event.}

\item{time_column}{The column name indicating the time in the users dataset}

\item{transient_phase}{An optional parameter to specify the transient phase (default is NULL).}

\item{eyeball_data}{A string ("yes" or "no") indicating whether to calculate initial decay parameters using simple maths. If "no" the user has to specify the initial decay parameters}

\item{plot_fitted}{A string ("yes" or "no") indicating whether to plot the fitted model.}

\item{decay_start}{The starting point (numeric) for the decay fitting. This should be specified in second.}

\item{y_Bas_user}{The user-defined baseline value (numeric) for the y variable.}

\item{A_p_user}{The user-defined amplitude value (numeric) for the decay model.}

\item{T_Dp_user}{The user-defined time delay value (numeric) for the decay model.}

\item{tau_p_user}{The user-defined time constant value (numeric) for the decay model.}
}
\value{
A data frame containing the fitted parameters for each variable in the input data.
}
\description{
This function processes data frames or CSV files to fit a single exponential decay model to specified variables.
It allows for manual or automatic calculation of decay parameters and provides options to visualize the fitted model.
Before calling this function, your data have to be smoothed, the function "filter_smooth" is suggested rather than the function "clean"
since no missing values are allowed.
}
\examples{
# Example usage with a list of data frames:
path_or_list <- filtered_smoothed
vars <- "TOI_1"
event_column <- "Comment"
start_event <- "EVNT13 "
transient_phase <- "on"
eyeball_data <- "yes"
plot_fitted <- "no"
decay_start <- 10
y_Bas_user <- 70
A_p_user <- 15
T_Dp_user <- 9
tau_p_user <- 5

decay <- single_decay(path_or_list, vars, event_column, start_event, transient_phase, eyeball_data, plot_fitted, decay_start, y_Bas_user, A_p_user, T_Dp_user, tau_p_user)
}
