% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_event.R
\name{slice_event}
\alias{slice_event}
\title{Slice Data Based on Events and Transient Phase}
\usage{
slice_event(
  file_list,
  period_col,
  events,
  sampling_freq,
  data_time,
  transient_phase
)
}
\arguments{
\item{file_list}{A list of data frames containing the datasets.}

\item{period_col}{The column name in the data frames containing the periods or events.}

\item{events}{A list of event identifiers to select rows around.}

\item{sampling_freq}{The sampling frequency of the NIRS data.}

\item{data_time}{Time duration (in seconds) of data to select around each event.}

\item{transient_phase}{Character indicating the transient phase: "on" or "off".}
}
\value{
A list of sliced data frames, where each element corresponds to a sliced dataset from each input data frame.
}
\description{
This function slices each dataset in a list of data frames based on specified events and transient phases.
It selects rows around the events depending on whether the transient phase is "on" or "off".
}
\examples{
# Example usage:
file_list <- cleaned_files
period_col <- "timestamp"
events <- list("event1", "event2")
sampling_freq <- 100  # Assuming 100 rows per second
data_time <- 10  # Select 10 seconds of data around each event
transient_phase <- "off"  # Select data before events
sliced_data <- slice_event(file_list, period_col, events, sampling_freq, data_time, transient_phase)
}
