% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_event.R
\name{slice_event}
\alias{slice_event}
\title{Slice Data Based on Events and Transient Phase}
\usage{
slice_event(
  file_list,
  period_col,
  events,
  sampling_freq,
  data_before,
  data_after,
  transient_phase
)
}
\arguments{
\item{file_list}{A list of data frames containing the datasets.}

\item{period_col}{The column name in the data frames containing the periods or events.}

\item{events}{A list of event identifiers to select rows around.}

\item{sampling_freq}{The sampling frequency of the NIRS data.}

\item{data_before}{Time duration (in seconds) of data to select before each event.}

\item{data_after}{Time duration (in seconds) of data to select after each event.}

\item{transient_phase}{Character indicating the transient phase: "heterogeneity" or "kinetics". If transient_phase is set to "heterogeneity" the data_after parameter is not used}
}
\value{
A list of sliced data frames, where each element corresponds to a sliced dataset from each input data frame.
}
\description{
This function slices each dataset in a list of data frames based on specified events.
Depending on the transient_phase the  data are sliced in a different way. For heterogeneity analyses the last seconds (user defined) before a specific period are sliced.
For kinetics analyses the data are sliced around the event depending on the user settings.
}
\examples{
# Example usage:
file_list <- cleaned_files
period_col <- "timestamp"
events <- list("event1", "event2")
sampling_freq <- 100  # Assuming 100 rows per second
data_before <- 10  # Select 10 seconds of data before each event
data_after <- 60 # Select 60 seconds of data after ach event
transient_phase <- "kinetics"
sliced_data <- slice_event(file_list, period_col, events, sampling_freq, data_before, data_after, transient_phase)
}
